pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "thinking-buildings"
dynamic = ["version"]
description = "Building security system using cameras and ML to detect people and animals"
readme = "README.md"
requires-python = ">=3.10"
license = "AGPL-3.0-only"
dependencies = [
    "opencv-python>=4.8.0",
    "pyyaml>=6.0",
    "plyer>=2.1.0",
    "onnxruntime>=1.16.0",
]

[project.optional-dependencies]
ultralytics = [
    "ultralytics>=8.3.0",
]
face = [
    "insightface>=0.7.3",
]
gpu = [
    "onnxruntime-gpu>=1.16.0",
]
dev = [
    "thinking-buildings[ultralytics,face]",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
    "ruff>=0.4.0",
]

[project.scripts]
thinking-buildings = "thinking_buildings.cli:main"

[tool.hatch.version]
source = "vcs"
fallback-version = "0.0.0-dev"

[tool.hatch.build.targets.wheel]
packages = ["thinking_buildings"]
